fix(native-transactions): relocate test fixture
